+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ## Release Notes
 
+### 2.5.3
+* Improved handling of untracked files
+* Added new Context entry `create versioned object`
+* Added command + button to edit configspec of current view
+
+### 2.4.3
+* Fix handling of network drive paths in the scm panel.
+
 ### 2.4.2
 * Security update